[distributed] add timeout for RPC futures, and ability to set timeout when initializing rpc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28392

Per #25531, we want to clean up futures when we detect that there are
failures/timeouts. As a first step, this diff adds timers to the future object,
provides functionality to check if a future is timed out, and allows
specification of the timeout when initializing rpc. A future diff will check for these timeouts and mark the future completed with an exception indicating that it has timed out.
ghstack-source-id: 93068370

Differential Revision: [D18025163](https://our.internmc.facebook.com/intern/diff/D18025163/)